decomposition, move business logic to workflow_runners.py
